DoFTools: move templates to dof_tools.templates.h
